protocol/Full_id: change Full_id union type
The union type is now uint64_t, for cases where an fd is >65535.
Truncation of the fd value would be a Bad Thing, so we update the
structure to hold both the 32-bit conn fd and the 16-bit request id.

However, there are 2 unused bytes now.

+++ include/mosh/fcgi/protocol/full_id.hpp
@@ -51,7 +51,7 @@ namespace protocol {
 		: fcgi_id(fcgi_id), fd(fd)
 		{ }
 
-		Full_id(uint32_t f)
+		Full_id(uint64_t f)
 		: full(f)
 		{ }
